Add correct Discord link to README.md
I tried to join the Discord channel from the link shared in the README but to no avail. I found the right link within MacForge but figured it might be a good idea to propose a change to the link here as well so others can find it. I'm not sure if this is the link you want to use, but it's the link that successfully brought me to the correct channel; feel free to change that if needed.